PR corrections
- Unified validations schema
- Cleaned login validations
- Secret key loaded from env
- Created sessions manager service
- Loaded user into request
- Unified tests files
- Created test data using factory
- Included expiration date in response
