+++ package.json
@@ -74,6 +74,7 @@
     "express-wolox-logger": "^2.0.0",
     "factory-girl": "^5.0.4",
     "jwt-simple": "^0.5.1",
+    "moment": "^2.29.1",
     "pg": "^7.4.1",
     "sequelize": "^5.21.5",
     "swagger-ui-express": "^4.0.7",
