+++ test/factory/users.js
@@ -1,0 +1,11 @@
+const { factory } = require('factory-girl');
+
+const { factoryWithCustomizedValue } = require('./factory_by_models');
+
+const modelName = 'User';
+
+factoryWithCustomizedValue(modelName, 'deletedAt', null);
+
+module.exports = {
+  create: userData => factory.create(modelName, userData)
+};
